requirements.txt
flask
python-dotenv
openai
chromadb
requests

